#物料/服务列表完成 #将pwd挪到home
