+++ app/home/views.py
@@ -1,9 +1,10 @@
 # -*- coding:utf-8 -*-
 from . import home
 from flask import render_template, session, redirect, request, url_for, flash
-from forms import LoginForm
-from app.models import User, Userlog
+from forms import LoginForm, PwdForm
+from app.models import User, Userlog, Oplog
 from app import db
+from werkzeug.security import generate_password_hash
 
 
 @home.route('/', methods=['GET'])
@@ -36,3 +37,35 @@ def logout():
     session.pop('user', None)
     session.pop('user_id', None)
     return redirect(url_for('home.login'))
+
+# 20180913 liuqq 修改密码
+@home.route('/user/pwd', methods=['GET', 'POST'])
+def pwd():
+    form = PwdForm()
+    if form.validate_on_submit():
+        # 验证密码
+        user = User.query.filter_by(id=session.get('user_id')).first()
+        if user.verify_password(form.old_pwd.data) != 1:
+            flash(u'旧密码输入错误', 'err')
+            return redirect(url_for('home.pwd'))
+        if form.new_pwd.data != form.re_pwd.data:
+            flash(u'您两次输入的密码不一致!', 'err')
+            return redirect(url_for('home.pwd'))
+        if form.new_pwd.data == form.old_pwd.data:
+            flash(u'新密码与旧密码一致！', 'err')
+            return redirect(url_for('home.pwd'))
+        new_pwd = generate_password_hash(form.new_pwd.data)
+        user.pwd = new_pwd
+        db.session.add(user)
+        # 20180917 jiangyu 增加修改密码日志
+        oplog = Oplog(
+            user_id=session['user_id'],
+            ip=request.remote_addr,
+            reason=u'修改密码:%s' % user.name
+        )
+        db.session.add(oplog)
+        db.session.commit()
+        flash(u'密码修改成功', 'ok')
+        return redirect(url_for('home.login'))
+    return render_template('home/pwd.html', form=form)
+
