+++ app/models.py
@@ -281,14 +281,24 @@ class Item(db.Model):
     id = db.Column(db.Integer, primary_key=True, autoincrement=True)
     # 名称
     name = db.Column(db.String(200), nullable=False)
-    # 类别
+    # 类别id
     cate_id = db.Column(db.Integer, db.ForeignKey('tb_category.id'))
+    # 类别(类别字段冗余，避免关联查询 type 0: item; 1: service)
+    type = db.Column(db.SmallInteger, default=0)
     # 销售价
     salesprice = db.Column(db.Float, default=0)
-    # 成本价
-    costprice = db.Column(db.Float, default=0)
     # 提成
     rewardprice = db.Column(db.Float, default=0)
+    # 成本价
+    costprice = db.Column(db.Float, default=0)
+    # 单位
+    unit = db.Column(db.String(40))
+    # 规格
+    standard = db.Column(db.String(100))
+    # 状态 (1有效；0无效)
+    valid = db.Column(db.SmallInteger, default=1)
+    # 备注
+    remarks = db.Column(db.Text)
     # 添加时间
     addtime = db.Column(db.DateTime, index=True, default=datetime.utcnow)
