+++ app/templates/base_sidebar.html
@@ -50,26 +50,22 @@
             <li><a href="#"><i class="fa fa-circle-o"></i> 会员列表</a></li>
         </ul>
     </li>
-    <li class="treeview">
-        <a href="#"><i class="fa fa-handshake-o"></i> <span> 供应商管理</span>
+    <li class="treeview" id="g-is">
+        <a href="#"><i class="fa fa-handshake-o"></i> <span> 商品/服务管理</span>
             <span class="pull-right-container">
-                            <i class="fa fa-angle-left pull-right"></i>
-                        </span>
+                <i class="fa fa-angle-left pull-right"></i>
+            </span>
         </a>
         <ul class="treeview-menu">
-            <li><a href="#"><i class="fa fa-circle-o"></i> 添加供应商</a></li>
-            <li><a href="#"><i class="fa fa-circle-o"></i> 供应商管理</a></li>
-        </ul>
-    </li>
-    <li class="treeview">
-        <a href="#"><i class="fa fa-link"></i> <span> 商品/服务管理</span>
-            <span class="pull-right-container">
-                            <i class="fa fa-angle-left pull-right"></i>
-                        </span>
-        </a>
-        <ul class="treeview-menu">
-            <li><a href="#"><i class="fa fa-circle-o"></i> 商品列表</a></li>
-            <li><a href="#"><i class="fa fa-circle-o"></i> 服务项目列表</a></li>
+            <li id="g-is-item">
+                <a href="#"><i class="fa fa-circle-o"></i> 商品列表</a>
+            </li>
+            <li id="g-is-service">
+                <a href="#"><i class="fa fa-circle-o"></i> 服务项目列表</a>
+            </li>
+            <li id="g-is-supplier">
+                <a href="#"><i class="fa fa-circle-o"></i> 供应商列表</a>
+            </li>
         </ul>
     </li>
     <li class="treeview" id="g-option">
